Clean up of static dependencies
