+++ production/skilljar-theme-v3.0/footer.mjs
@@ -1,12 +1,18 @@
 import { el, sanitizeUrl } from "./meta.mjs";
 
+// static imports
+import { footerData } from "../data/footer.mjs";
+
 /**
  * Generates the footer HTML and appends it to the specified container.
  * @param {Object} data - The footer data object.
  * @param {string} containerId - The ID of the container to append the footer to.
  * @returns {void}
  */
-export function generateFooter(data, containerId = "footer-container") {
+export function generateFooter(
+  data = footerData,
+  containerId = "footer-container",
+) {
   const container = document.getElementById(containerId);
   if (!container) return;
   container.innerHTML = "";
@@ -26,7 +32,7 @@ export function generateFooter(data, containerId = "footer-container") {
             acc.push(el("span", { text: " | " }));
           }
           return acc;
-        }, [])
+        }, []),
     ),
   ];
 
@@ -72,10 +78,10 @@ export function generateFooter(data, containerId = "footer-container") {
                     "M8.4 8.4L8.4 5.6L5.6 5.6L5.6 8.4L8.4 8.4Z",
                     "M14 8.4L14 5.6L11.2 5.6L11.2 8.4L14 8.4Z",
                     "M2.8 8.4L2.8 5.6L-2.15213e-06 5.6L-2.27452e-06 8.4L2.8 8.4Z",
-                  ].map((d) => el("path", { d }))
+                  ].map((d) => el("path", { d })),
                 ),
-              ]
-            )
+              ],
+            ),
           )
         : undefined,
 
@@ -99,8 +105,8 @@ export function generateFooter(data, containerId = "footer-container") {
             href: sanitizeUrl(link.href),
             target: "_blank",
             text: link.label,
-          })
-        )
+          }),
+        ),
       );
       colDiv.appendChild(gDiv);
     });
@@ -134,8 +140,8 @@ export function generateFooter(data, containerId = "footer-container") {
               href: sanitizeUrl(s.href),
               target: "_blank",
               innerHTML: s.svg,
-            })
-          )
+            }),
+          ),
         )
       : undefined,
   ]);
@@ -145,7 +151,7 @@ export function generateFooter(data, containerId = "footer-container") {
   const copyrightFooter = el(
     "div",
     { className: "footer-copyright" },
-    copyrightContent
+    copyrightContent,
   );
   container.appendChild(copyrightFooter);
 }
