+++ production/skilljar-theme-v3.0/static.mjs
@@ -57,123 +57,8 @@ export const CONFIG = {
     },
   },
   codeTheme: "min-light", // Shiki theme to use for code blocks
-  icons: {
-    checkbox: {
-      paths: [
-        "M8.22948 14.021L5.02148 10.792L5.75048 10.042L8.22948 12.5L14.2505 6.5L14.9795 7.271L8.22948 14.021Z",
-      ],
-    },
-    copy: {
-      attrs: {
-        height: "20",
-        width: "21",
-      },
-      paths: [
-        "M5.12597 18.0835C4.75064 18.0835 4.43464 17.9548 4.17797 17.6975C3.9213 17.4408 3.79297 17.1248 3.79297 16.7495V6.3335H4.87597V16.7495C4.87597 16.8195 4.9003 16.8785 4.94897 16.9265C4.99764 16.9752 5.05664 16.9995 5.12597 16.9995H13.543V18.0835H5.12597ZM7.70897 15.4995C7.3343 15.4995 7.0183 15.3712 6.76097 15.1145C6.5043 14.8572 6.37597 14.5412 6.37597 14.1665V4.6875C6.37597 4.31216 6.5043 3.99283 6.76097 3.7295C7.0183 3.4655 7.3343 3.3335 7.70897 3.3335H15.189C15.5636 3.3335 15.883 3.4655 16.147 3.7295C16.411 3.99283 16.543 4.31216 16.543 4.6875V14.1665C16.543 14.5412 16.411 14.8572 16.147 15.1145C15.883 15.3712 15.5636 15.4995 15.189 15.4995H7.70897ZM7.70897 14.4165H15.189C15.2583 14.4165 15.3206 14.3922 15.376 14.3435C15.4313 14.2948 15.459 14.2358 15.459 14.1665V4.6875C15.459 4.61816 15.4313 4.5555 15.376 4.4995C15.3206 4.44416 15.2583 4.4165 15.189 4.4165H7.70897C7.63964 4.4165 7.58064 4.44416 7.53197 4.4995C7.4833 4.5555 7.45897 4.61816 7.45897 4.6875V14.1665C7.45897 14.2358 7.4833 14.2948 7.53197 14.3435C7.58064 14.3922 7.63964 14.4165 7.70897 14.4165Z",
-      ],
-    },
-    bookmark: {
-      attrs: {
-        version: "1.1",
-        "xmlns:xlink": "http://www.w3.org/1999/xlink",
-        viewBox: "0 0 963.4 963.4",
-        xmlSpace: "preserve",
-        width: "24",
-        height: "24",
-      },
-      paths: [
-        "M114.3,59.2H69.7c-33.1,0-60,26.9-60,60V903.4c0,33.1,26.9,60,60,60h824c33.1,0,60-26.9,60-60V119.2c0-33.1-26.9-60-60-60   H568.9c0,0.3,0,0.5,0,0.8v420.3c0,25.6-10.2,49.2-28.8,66.5c-17,15.799-39.2,24.5-62.4,24.5c-12.4,0-24.4-2.5-35.7-7.301   c-11.899-5.1-22.399-12.6-31.2-22.301L341.601,466.1l-69.2,75.599C263.5,551.4,253,558.9,241.2,564   c-11.3,4.9-23.3,7.301-35.7,7.301c-23.2,0-45.4-8.701-62.4-24.5c-18.6-17.301-28.8-40.9-28.8-66.5V60   C114.3,59.7,114.3,59.4,114.3,59.2z",
-        "M228.2,501.1l90.6-99.1c6.101-6.699,14.5-10.1,22.9-10.1s16.7,3.4,22.9,10.1l90.6,99.1c6.4,7,14.6,10.1,22.6,10.1   c15.9,0,31.301-12.299,31.301-31.099V60c0-0.3,0-0.5,0-0.8C508.7,26.4,482,0,449.101,0H234.3c-32.9,0-59.6,26.4-60,59.2   c0,0.3,0,0.5,0,0.8v420.3c0,18.799,15.3,31.1,31.3,31.1C213.6,511.301,221.7,508.199,228.2,501.1z",
-      ],
-    },
-    burger: {
-      attrs: {
-        width: "24",
-        height: "24",
-        viewBox: "0 0 24 24",
-      },
-      paths: ["M4 4h16v2H4V4Zm0 7h16v2H4v-2Zm0 7h16v2H4v-2Z"],
-    },
-    chainguard: {
-      attrs: {
-        width: "30",
-        height: "27",
-        viewBox: "0 0 30 27",
-      },
-      paths: [
-        "M24.7199 16.1854C24.9569 15.368 25.0857 14.4724 25.0857 13.4982C25.0857 7.40502 20.047 0.546799 15.0146 0.546799C9.98217 0.546799 4.94345 7.40502 4.94345 13.4982C4.94345 14.8267 5.18298 16.0092 5.60976 17.0461L2.58476 16.876C1.35268 16.8067 0.217578 17.6851 0.409743 18.9041C0.484274 19.3769 0.626315 19.8608 0.879992 20.2788C0.0236265 20.8812 -0.326192 21.9885 0.367699 22.8564C1.06428 23.7277 2.08704 24.5305 3.49093 24.5305C5.01364 24.5305 5.93005 24.137 6.48659 23.6428C6.52721 23.8586 6.61101 24.0711 6.7433 24.2719C7.42673 25.3095 8.55862 26.4501 10.232 26.4501C13.0786 26.4501 13.3961 24.6622 13.5554 23.765C13.5679 23.6948 13.5794 23.6301 13.591 23.572L15.4933 22.6207L17.3956 23.572C17.4072 23.63 17.4187 23.6947 17.4312 23.7648L17.4312 23.765C17.5905 24.6622 17.908 26.4501 20.7546 26.4501C22.428 26.4501 23.5599 25.3095 24.2433 24.2719C24.2702 24.2311 24.2951 24.1898 24.318 24.1481C24.8542 24.38 25.5641 24.5305 26.506 24.5305C27.9099 24.5305 28.9327 23.7277 29.6292 22.8564C30.4824 21.7893 29.7577 20.3602 28.4536 19.9528L28.0227 19.8181C29.2881 19.1624 29.4743 17.9255 29.3387 16.8418C29.1856 15.6172 27.8516 15.0879 26.687 15.496L24.7199 16.1854ZM15.4951 22.603C15.494 22.603 15.4929 22.603 15.4917 22.6031L15.4933 22.6039L15.4951 22.603Z",
-      ],
-    },
-    lightning: {
-      attrs: {
-        width: "24",
-        height: "24",
-        viewBox: "0 0 24 24",
-      },
-      paths: ["M13.5 2 5 13h6l-1.2 9L19 11h-6l.5-9z"],
-    },
-    star: {
-      attrs: {
-        width: "24",
-        height: "24",
-        viewBox: "0 0 24 24",
-      },
-      paths: ["M12 2 15 8 22 9 17 14l1 7-6-3-6 3 1-7-5-5 7-1 3-6Z"],
-    },
-    shield: {
-      attrs: {
-        width: "21",
-        height: "25",
-        viewBox: "0 0 21 25",
-      },
-      paths: [
-        "M10.5 0.309998L0.748047 3.967V11.3907C0.748047 17.5467 4.90484 23.2881 10.5 24.69C16.0953 23.2881 20.252 17.5467 20.252 11.3907V3.967L10.5 0.309998ZM17.814 11.3907C17.814 16.2667 14.7056 20.777 10.5 22.1545C6.2945 20.777 3.18605 16.2789 3.18605 11.3907V5.66141L10.5 2.91866L17.814 5.66141V11.3907Z",
-      ],
-    },
-    frames: {
-      attrs: {
-        width: "29",
-        height: "29",
-        viewBox: "0 0 29 29",
-      },
-      paths: [
-        "M14.9488 28.3542C14.2121 28.3542 13.5886 28.099 13.0782 27.5885C12.5677 27.0781 12.3125 26.4546 12.3125 25.7179V14.9488C12.3125 14.2121 12.5677 13.5885 13.0782 13.0781C13.5886 12.5677 14.2121 12.3125 14.9488 12.3125H25.7179C26.4546 12.3125 27.0782 12.5677 27.5886 13.0781C28.099 13.5885 28.3542 14.2121 28.3542 14.9488V25.7179C28.3542 26.4546 28.099 27.0781 27.5886 27.5885C27.0782 28.099 26.4546 28.3542 25.7179 28.3542H14.9488ZM14.9488 26.1667H25.7179C25.8302 26.1667 25.933 26.1199 26.0263 26.0263C26.1199 25.933 26.1667 25.8302 26.1667 25.7179V14.9488C26.1667 14.8365 26.1199 14.7337 26.0263 14.6404C25.933 14.5468 25.8302 14.5 25.7179 14.5H14.9488C14.8366 14.5 14.7337 14.5468 14.6404 14.6404C14.5468 14.7337 14.5 14.8365 14.5 14.9488V25.7179C14.5 25.8302 14.5468 25.933 14.6404 26.0263C14.7337 26.1199 14.8366 26.1667 14.9488 26.1667ZM6.47921 22.6889V9.11546C6.47921 8.37876 6.73442 7.7552 7.24483 7.24479C7.75525 6.73437 8.37881 6.47916 9.11551 6.47916H22.6889V8.66666H9.11551C9.00322 8.66666 8.90041 8.71345 8.80707 8.80703C8.7135 8.90036 8.66671 9.00317 8.66671 9.11546V22.6889H6.47921ZM0.645874 16.8556V3.28213C0.645874 2.54543 0.901082 1.92187 1.4115 1.41145C1.92192 0.901037 2.54547 0.645828 3.28218 0.645828H16.8556V2.83333H3.28218C3.16988 2.83333 3.06707 2.88012 2.97374 2.97369C2.88016 3.06703 2.83337 3.16984 2.83337 3.28213V16.8556H0.645874Z",
-      ],
-    },
-    classBurger: {
-      attrs: {
-        width: "24",
-        height: "24",
-        viewBox: "0 0 24 24",
-      },
-      paths: ["M4 6h16v2H4V6Zm0 5h16v2H4v-2Zm0 5h16v2H4v-2Z"],
-    },
-    rightArrow: {
-      attrs: {
-        viewBox: "0 0 14 14",
-        width: "20",
-        height: "20",
-        fill: "white",
-        ariaHidden: "true",
-      },
-      paths: [
-        "M11.2 5.6L11.2 2.8L8.4 2.8L8.4 5.6L11.2 5.6Z",
-        "M11.2 11.2L11.2 8.4L8.4 8.4L8.4 11.2L11.2 11.2Z",
-        "M8.4 2.8L8.4 -2.44784e-07L5.6 -3.67176e-07L5.6 2.8L8.4 2.8Z",
-        "M8.4 14L8.4 11.2L5.6 11.2L5.6 14L8.4 14Z",
-        "M8.4 8.4L8.4 5.6L5.6 5.6L5.6 8.4L8.4 8.4Z",
-        "M14 8.4L14 5.6L11.2 5.6L11.2 8.4L14 8.4Z",
-        "M2.8 8.4L2.8 5.6L-2.15213e-06 5.6L-2.27452e-06 8.4L2.8 8.4Z",
-      ],
-    },
-  },
 };
 
-export const bannerSVGs = {
-  left: `<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 303 303" aria-hidden="true"><path fill="#fff" stroke="#ededed" d="M201.6.361H252v50.4h-50.4zm0 50.4H252v50.4h-50.4z"/><path fill="#f8f6fe" stroke="#ededed" d="M201.6 101.161H252v50.4h-50.4z"/><path fill="#6226fb" d="M201.6 151.561H252v50.4h-50.4z"/><path fill="#fff" stroke="#ededed" d="M201.6 201.96H252v50.4h-50.4z"/><path fill="#f8f6fe" stroke="#ededed" d="M201.6 252.361H252v50.4h-50.4z"/><path fill="#6226fb" d="M151.2.361h50.4v50.4h-50.4z"/><path fill="#f1ecfe" stroke="#ededed" d="M151.2 50.761h50.4v50.4h-50.4z"/><path fill="#fff" stroke="#ededed" d="M151.2 101.161h50.4v50.4h-50.4zm0 50.4h50.4v50.4h-50.4z"/><path fill="#fff" stroke="#ededed" d="M151.2 201.96h50.4v50.4h-50.4zm0 50.401h50.4v50.4h-50.4z"/><path fill="#f8f6fe" stroke="#ededed" d="M252 .361h50.4v50.4H252z"/><path fill="#fff" stroke="#ededed" d="M252 50.761h50.4v50.4H252z"/><path fill="#6226fb" d="M252 101.161h50.4v50.4H252z"/><path fill="#fff" stroke="#ededed" d="M252 151.561h50.4v50.4H252z"/><path fill="#f1ecfe" stroke="#ededed" d="M252 201.96h50.4v50.4H252z"/><path fill="#fff" stroke="#ededed" d="M252 252.361h50.4v50.4H252zM50.4.361h50.4v50.4H50.4zm0 50.4h50.4v50.4H50.4z"/><path fill="#6226fb" d="M50.4 101.161h50.4v50.4H50.4z"/><path fill="#f1ecfe" stroke="#ededed" d="M50.4 151.561h50.4v50.4H50.4z"/><path fill="#fff" stroke="#ededed" d="M50.4 201.96h50.4v50.4H50.4z"/><path fill="#6226fb" d="M50.4 252.361h50.4v50.4H50.4zM0 .361h50.4v50.4H0z"/><path fill="#f8f6fe" stroke="#ededed" d="M0 50.761h50.4v50.4H0z"/><path fill="#fff" stroke="#ededed" d="M0 101.161h50.4v50.4H0zm0 50.4h50.4v50.4H0z"/><path fill="#fff" stroke="#ededed" d="M0 201.96h50.4v50.4H0z"/><path fill="#f1ecfe" stroke="#ededed" d="M0 252.361h50.4v50.4H0z"/><path fill="#fff" stroke="#ededed" d="M100.8.361h50.4v50.4h-50.4zm0 50.4h50.4v50.4h-50.4z"/><path fill="#f8f6fe" d="M100.8 101.161h50.4v50.4h-50.4z"/><path fill="#fff" stroke="#ededed" d="M100.8 151.561h50.4v50.4h-50.4z"/><path fill="#6226fb" d="M100.8 201.96h50.4v50.4h-50.4z"/><path fill="#fff" stroke="#ededed" d="M100.8 252.361h50.4v50.4h-50.4z"/></svg>`,
-  right: `<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 303 303" aria-hidden="true"><path fill="#fff" stroke="#ededed" d="M101.3.36H50.9v50.4h50.4zm0 50.4H50.9v50.4h50.4z"/><path fill="#fef5fe" stroke="#ededed" d="M101.3 101.161H50.9v50.4h50.4z"/><path fill="#fd2bf2" d="M101.3 151.561H50.9v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M101.3 201.96H50.9v50.4h50.4z"/><path fill="#fef5fe" stroke="#ededed" d="M101.3 252.361H50.9v50.4h50.4z"/><path fill="#fd2bf2" d="M151.7.36h-50.4v50.4h50.4z"/><path fill="#fddffc" stroke="#ededed" d="M151.7 50.76h-50.4v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M151.7 101.161h-50.4v50.4h50.4zm0 50.4h-50.4v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M151.7 201.96h-50.4v50.4h50.4zm0 50.401h-50.4v50.4h50.4z"/><path fill="#fef5fe" stroke="#ededed" d="M50.9.36H.5v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M50.9 50.76H.5v50.4h50.4z"/><path fill="#fd2bf2" d="M50.9 101.161H.5v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M50.9 151.561H.5v50.4h50.4z"/><path fill="#fddffc" stroke="#ededed" d="M50.9 201.96H.5v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M50.9 252.361H.5v50.4h50.4zM252.5.36h-50.4v50.4h50.4zm0 50.4h-50.4v50.4h50.4z"/><path fill="#fd2bf2" d="M252.5 101.161h-50.4v50.4h50.4z"/><path fill="#fddffc" stroke="#ededed" d="M252.5 151.561h-50.4v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M252.5 201.96h-50.4v50.4h50.4z"/><path fill="#fd2bf2" d="M252.5 252.361h-50.4v50.4h50.4zM302.9.36h-50.4v50.4h50.4z"/><path fill="#fef5fe" stroke="#ededed" d="M302.9 50.76h-50.4v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M302.9 101.161h-50.4v50.4h50.4zm0 50.4h-50.4v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M302.9 201.96h-50.4v50.4h50.4z"/><path fill="#fddffc" stroke="#ededed" d="M302.9 252.361h-50.4v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M202.1.36h-50.4v50.4h50.4zm0 50.4h-50.4v50.4h50.4z"/><path fill="#fef5fe" d="M202.1 101.161h-50.4v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M202.1 151.561h-50.4v50.4h50.4z"/><path fill="#fd2bf2" d="M202.1 201.96h-50.4v50.4h50.4z"/><path fill="#fff" stroke="#ededed" d="M202.1 252.361h-50.4v50.4h50.4z"/></svg>`,
-  arrow: `<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 16 16" aria-hidden="true"><path d="M12.8 6.4V3.2H9.6v3.2h3.2ZM12.8 12.8V9.6H9.6v3.2h3.2ZM9.6 3.2V0H6.4v3.2h3.2ZM9.6 16v-3.2H6.4V16h3.2Z"/><path d="M9.6 9.6V6.4H6.4v3.2h3.2ZM16 9.6V6.4h-3.2v3.2H16ZM3.2 9.6V6.4H0v3.2h3.2Z"/></svg>`,
-};
-
+export { icons, bannerSVGs } from "../data/graphics.mjs";
 export { footerData } from "../data/footer.mjs";
 export { pathSections } from "../data/path-sections.mjs";
