+++ production/theme.mjs
@@ -15,7 +15,6 @@
 
 import { Q, el, remove } from "./skilljar-theme-v3.0/meta.mjs";
 import { CG } from "./skilljar-theme-v3.0/CG.mjs";
-import { footerData } from "./skilljar-theme-v3.0/static.mjs";
 import { animateCompletion } from "./skilljar-theme-v3.0/course-completion.mjs";
 import { generateFooter } from "./skilljar-theme-v3.0/footer.mjs";
 import { hide, showBody } from "./skilljar-theme-v3.0/styling.mjs";
@@ -31,7 +30,7 @@ document.addEventListener("DOMContentLoaded", () => {
   CG.dom.bodyHeader.classList.add("headers");
   hide("#ep-footer"); // hide Skilljar footer
 
-  generateFooter(footerData);
+  generateFooter();
   window.animateCompletion = animateCompletion;
 
   if (CG.page.isLesson)
